Move TUI keybind helpers into tui-core
