+++ crates/jcode-tui-core/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod keybind;
 pub mod stream_buffer;
 
 pub use stream_buffer::{StreamBuffer, StreamBufferMemoryProfile};
